tests updating, split tests to different classes as needed - better for parallel run.
