+++ GitHubProjectFolder/helpers/Categories.cs
@@ -1,10 +1,11 @@
-﻿namespace GitHub.config
+﻿namespace GitHub.helpers
 {
     public static class Categories
     {
         public const string SanityTest = "SanityTest";
         public const string RegressionTest = "RegressionTest";
         public const string Login = "Login";
-        public const string Registration = "Registration";
+        public const string Repository = "Repository";
+        public const string Issue = "Issue";
     }
 }
